Adds Environment Canada as a weather source
Also makes a number of changes to Weather in general:

- Make use of Mojo::AsyncAwait where we used to use promises and
callbacks.
- Now callbacks for weather are optional, you can also use await.
- Weather command looks at the address returned by Google Maps and
determines if it is in Canada. If so, it uses Environment Canada as
the data source
- No error handling yet for Canadian cities which are not listed in
Environment Canada's repo. Eventually this should default back to Dark
Sky.
- Link for radar and forecast always uses Dark Sky still.
- New icons for Environment Canada, hosted by them
- Offloaded some of the json datastructure responsibilities to the
componenets, namely ensuring things like temperature and wind speed are
in both units, and ensuring icon URL and icon Emotes exist.
- Lots of debug output is still enabled, this is still WIP.