!duolingo extra and !duolingo extra <num> are now a thing
Duolingo made changes to their crown system at some point in the past
and decided to just carry forward the old Crowns value and add it on to
the new Crowns value. The problem is this value is only shown via the
API to the authenticated user. This means the bot can only see its own
"extraCrowns" value, not anyone else's. This leads to a discrepancy
between the number of crowns reported on the Duolingo website / app, and
the number of crowns reported by the bot.

So this change introduces the ability to tell the bot how many crowns to
add to their total so the bot's math lines up with the site's math.

There is no way to validate the value, it is just limited to a range of
0-255.

!duo extra - Display your current extraCrowns value
!duo extra <num> - Set your extraCrowns value

This value is tied to the current language you are learning, and will
track a separate extraCrowns value for each language you use.